+++ src/lib/components/consignments/__stories__/consignments-todo.story.tsx
@@ -4,10 +4,19 @@ import * as React from "react"
 import TODO from "../components/artwork-consignment-todo"
 import { ConsignmentMetadata, ConsignmentSetup } from "../index"
 
+export const name = "Consignments - TODO"
+export const component = TODO
+
+interface States {
+  [name: string]: ConsignmentSetup;
+}
+
 const withArtist: ConsignmentSetup = { artist: { name: "Glenn Brown" } }
 const withOnePhoto: ConsignmentSetup = {
   ...withArtist,
-  photos: ["https://d32dm0rphc51dk.cloudfront.net/VFiyokWNcBZNlfglZND_3g/small_square.jpg"],
+  photos: [
+    "https://d32dm0rphc51dk.cloudfront.net/VFiyokWNcBZNlfglZND_3g/small_square.jpg",
+  ],
 }
 
 const withPhotos: ConsignmentSetup = {
@@ -45,23 +54,17 @@ const withProvenance: ConsignmentSetup = {
   provenance: "This work has seen many hands.",
 }
 
-export const allStates = {
-  withArtist,
-  withPhotos,
-  withMetadata,
-  withLocation,
-  withProvenance,
-}
+const longProv =
+  "This is a long long long run on sentence that should break correctly."
 
-const longProv = "This is a long long long run on sentence that should break correctly."
-
-storiesOf("Consignments - TODO")
-  .add("Empty", () => <TODO />)
-  .add("With Artist", () => <TODO {...withArtist} />)
-  .add("With Photos", () => <TODO {...withPhotos} />)
-  .add("With Metadata", () => <TODO {...withMetadata} />)
-  .add("With Location", () => <TODO {...withLocation} />)
-  .add("With Provenance", () => <TODO {...withProvenance} />)
-  .add("With Just Metadata", () => <TODO metadata={metadata} />)
-  .add("With One Photo", () => <TODO {...withOnePhoto} />)
-  .add("With Long Provenance", () => <TODO provenance={longProv} />)
+export const allStates: States[] = [
+  { "Empty Metadata": {} },
+  { "With Artist": withArtist },
+  { "With Photos": withPhotos },
+  { "With Metadata": withMetadata },
+  { "With Location": withMetadata },
+  { "With Provenance": withProvenance },
+  { "With Just Metadata": { metadata } },
+  { "With One": withOnePhoto },
+  { "With Long": { provenance: longProv } },
+]