+++ src/lib/components/consignments/components/__tests__/bottom-aligned-button-tests.tsx
@@ -1,0 +1,4 @@
+import * as BottomAligned from "../../__stories__/consignments-bottom-aligned.story"
+import storyRunner from "./runner"
+
+storyRunner("Bottom-Aligned states: ", BottomAligned)